Extract clause-relation graph and punctuation segmentation
